🚚 move: 타임테이블 시간 정보 함수 분리
