+++ src/utils/realtimeBar.js
@@ -13,4 +13,16 @@ export const realtimeBar = (festaDate) =>{
   return diffMin;
 };
 
+// ---- 타임테이블 표 생성을 위한 시간 정보 ----
+export const timeSlot = () =>{
+  const timeSlots = [];
 
+  for (let hour = 13; hour < 22; hour++) {
+    timeSlots.push(`${hour}:00`);
+    timeSlots.push(`${hour}:30`);
+  }
+
+  timeSlots.push("22:00");
+
+  return timeSlots
+}
